commented out some of the code on vendor-router. Should work better now. Changed sone of the migrations to decimal instead of float.

+++ data/migrations/20200524083851_vendors_table.js
@@ -58,7 +58,7 @@ exports.up = function (knex) {
       tbl.text("product_category", 255);
       tbl.text("diet_category", 255);
       tbl.text("unit", 255);
-      tbl.float("price").notNullable();
+      tbl.decimal("price").notNullable();
     })
 
     .createTable("vendor_customer_map", (tbl) => {
